requirements.txt
APIFlask==2.1.1
Flask_SQLAlchemy==3.0.2
marshmallow==3.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
SQLAlchemy_Utils==0.41.2
Werkzeug==3.0.1
